# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Main API for the ProjectXiel website",
        "title": "ProjectXiel API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "fiber@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/post/{slug}": {
            "get": {
                "description": "get post by slug",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Post"
                ],
                "summary": "Get a single post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Post Slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.Post"
                        }
                    }
                }
            }
        },
        "/posts": {
            "get": {
                "description": "get posts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Post"
                ],
                "summary": "Get all posts",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Posts Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Posts page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.Post"
                        }
                    }
                }
            }
        },
        "/youtube/videos": {
            "get": {
                "description": "Fetches a list of YouTube videos from a specific channel.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "YouTube"
                ],
                "summary": "Fetch YouTube videos",
                "responses": {
                    "200": {
                        "description": "A list of YouTube videos",
                        "schema": {
                            "$ref": "#/definitions/data.YouTubeResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "data.PageInfo": {
            "type": "object",
            "properties": {
                "resultsPerPage": {
                    "type": "integer"
                },
                "totalResults": {
                    "type": "integer"
                }
            }
        },
        "data.Post": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "date_created": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "data.Thumbnail": {
            "type": "object",
            "properties": {
                "height": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
        "data.YouTubeID": {
            "type": "object",
            "properties": {
                "channelId": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "videoId": {
                    "type": "string"
                }
            }
        },
        "data.YouTubeItem": {
            "type": "object",
            "properties": {
                "etag": {
                    "type": "string"
                },
                "id": {
                    "$ref": "#/definitions/data.YouTubeID"
                },
                "kind": {
                    "type": "string"
                },
                "snippet": {
                    "$ref": "#/definitions/data.YouTubeSnippet"
                }
            }
        },
        "data.YouTubeResponse": {
            "type": "object",
            "properties": {
                "etag": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/data.YouTubeItem"
                    }
                },
                "kind": {
                    "type": "string"
                },
                "pageInfo": {
                    "$ref": "#/definitions/data.PageInfo"
                },
                "regionCode": {
                    "type": "string"
                }
            }
        },
        "data.YouTubeSnippet": {
            "type": "object",
            "properties": {
                "channelId": {
                    "type": "string"
                },
                "channelTitle": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "liveBroadcastContent": {
                    "type": "string"
                },
                "publishTime": {
                    "type": "string"
                },
                "publishedAt": {
                    "type": "string"
                },
                "thumbnails": {
                    "$ref": "#/definitions/data.YouTubeThumbnails"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "data.YouTubeThumbnails": {
            "type": "object",
            "properties": {
                "default": {
                    "$ref": "#/definitions/data.Thumbnail"
                },
                "high": {
                    "$ref": "#/definitions/data.Thumbnail"
                },
                "medium": {
                    "$ref": "#/definitions/data.Thumbnail"
                }
            }
        }
    }
}